pyproject.toml
[tool.poetry]
name = "pbsv2"
version = "1.0.0"
description = "Bushfire Management System"
authors = ["DBCA OIM <asi@dbca.wa.gov.au>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.10"
Django = "5.0.9"
djangorestframework = "^3.14"
dj-database-url = "^2.1"
psycopg2 = "^2.9"
gunicorn = "^23.0"
whitenoise = "^6.5"
dbca-utils = {git = "https://github.com/dbca-wa/dbca-utils.git", rev = "1.1.1"}
webtemplate_dbca = {git = "https://github.com/dbca-wa/webtemplate2.git", rev = "master"}
wagov_utils = {git = "https://github.com/dbca-wa/wagov_utils.git"}
drf-spectacular = "^0.26"
django-filter = "^23.3"
python-dateutil = "^2.8"
django-reversion = "^5.0"
django-reversion-rest-framework = "^1.1"
django-cron = "^0.6"
django-cors-headers = "^4.2"
py7zr = "^0.20"
httpx = "^0.25"
owslib = "^0.29"
numpy = "^1.25"
django-extensions = "^3.2"
django-model-utils = "^4.3"
coverage = "^7.3"
django-coverage-plugin = "^3.1"
setuptools = "^70.3"
django-protected-media = "^1.0"
selenium = "^4.14"
django-selenium = "^0.9"
django-storages = "^1.14"
sentry-sdk = {extras = ["django"], version = "^1.32"}
django-nested-admin = "^4.0.2"
drf-standardized-errors = "^0.12.6"
djangorestframework-datatables = "^0.7.0"
djangorestframework-gis = "^1.0"
django-debug-toolbar = "^4.3.0"
appmonitor_client = {git = "https://github.com/dbca-wa/appmonitor_client.git", rev="main"}
python-decouple = "^3.8"
tomli = "^2.0.1"
django-dirtyfields = "^1.9.2"
django-mail-queue = "^3.2.5"
drf-orjson-renderer = "^1.7.2"

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.2"
ipython = "^8.3"
ipdb = "^0.13"
pylint = "^3.0"
flake8 = "^6.1"
autopep8 = "^2.0"
black = "^23.9"
mypy = "^1.6.1"
django-stubs = "^4.2"
django-stubs-ext = "^4.2"
djangorestframework-stubs = "^3.14"
pyparsing = "^3.1.1"
pydot = "^1.4.2"

[build-system]
requires = ["poetry-core>=1.6.1"]
build-backend = "poetry.core.masonry.api"
